+++ testutil/driver.go
@@ -15,7 +15,6 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Operations"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/cluster"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver"
 )
 
 var ErrNotImplemented = errors.New("testutil: not implemented")
@@ -171,13 +170,6 @@ func (db *db) Invoke(ctx context.Context, method string, args interface{}, reply
 	if db.onInvoke == nil {
 		return fmt.Errorf("db.onInvoke() not implemented")
 	}
-	defer func() {
-		if err == nil {
-			if apply, ok := driver.ContextCallInfo(ctx); ok && apply != nil {
-				apply(db)
-			}
-		}
-	}()
 	return db.onInvoke(ctx, method, args, reply, opts...)
 }
 
@@ -185,17 +177,10 @@ func (db *db) NewStream(ctx context.Context, desc *grpc.StreamDesc, method strin
 	if db.onNewStream == nil {
 		return nil, fmt.Errorf("db.onNewStream() not implemented")
 	}
-	defer func() {
-		if err == nil {
-			if apply, ok := driver.ContextCallInfo(ctx); ok && apply != nil {
-				apply(db)
-			}
-		}
-	}()
 	return db.onNewStream(ctx, desc, method, opts...)
 }
 
-func (db *db) Get(context.Context) (conn cluster.ClientConnInterface, err error) {
+func (db *db) Get(context.Context) (conn grpc.ClientConnInterface, err error) {
 	cc := &clientConn{
 		onInvoke:    db.onInvoke,
 		onNewStream: db.onNewStream,